# perf(fslib): skip `lstat` if parent didn't exist