--- a/.yarn/versions/c07043ca.yml
+++ b/.yarn/versions/c07043ca.yml
@@ -1,9 +1,12 @@
 releases:
   "@yarnpkg/cli": patch
+  "@yarnpkg/core": patch
   "@yarnpkg/fslib": patch
   "@yarnpkg/plugin-pnp": patch
+  "@yarnpkg/pnp": patch
 
 declined:
+  - "@yarnpkg/esbuild-plugin-pnp"
   - "@yarnpkg/plugin-compat"
   - "@yarnpkg/plugin-constraints"
   - "@yarnpkg/plugin-dlx"
@@ -28,11 +31,9 @@ declined:
   - "@yarnpkg/plugin-workspace-tools"
   - vscode-zipfs
   - "@yarnpkg/builder"
-  - "@yarnpkg/core"
   - "@yarnpkg/doctor"
   - "@yarnpkg/json-proxy"
   - "@yarnpkg/nm"
-  - "@yarnpkg/pnp"
   - "@yarnpkg/pnpify"
   - "@yarnpkg/sdks"
   - "@yarnpkg/shell"