--- a/packages/yarnpkg-fslib/sources/algorithms/copyPromise.ts
+++ b/packages/yarnpkg-fslib/sources/algorithms/copyPromise.ts
@@ -128,20 +128,19 @@ async function copyFolder<P1 extends Path, P2 extends Path>(prelayout: Operation
 
   const entries = await sourceFs.readdirPromise(source);
 
-  const copyImplOpts: CopyImplOptions = {
-    ...opts,
-    didParentExist: !!destinationStat,
-  };
+  const nextOpts = opts.didParentExist && !destinationStat
+    ? {...opts, didParentExist: false}
+    : opts;
 
   if (opts.stableSort) {
     for (const entry of entries.sort()) {
-      if (await copyImpl(prelayout, postlayout, updateTime, destinationFs, destinationFs.pathUtils.join(destination, entry), sourceFs, sourceFs.pathUtils.join(source, entry), copyImplOpts)) {
+      if (await copyImpl(prelayout, postlayout, updateTime, destinationFs, destinationFs.pathUtils.join(destination, entry), sourceFs, sourceFs.pathUtils.join(source, entry), nextOpts)) {
         updated = true;
       }
     }
   } else {
     const entriesUpdateStatus = await Promise.all(entries.map(async entry => {
-      await copyImpl(prelayout, postlayout, updateTime, destinationFs, destinationFs.pathUtils.join(destination, entry), sourceFs, sourceFs.pathUtils.join(source, entry), copyImplOpts);
+      await copyImpl(prelayout, postlayout, updateTime, destinationFs, destinationFs.pathUtils.join(destination, entry), sourceFs, sourceFs.pathUtils.join(source, entry), nextOpts);
     }));
 
     if (entriesUpdateStatus.some(status => status)) {