**What's the problem this PR addresses?**

`@yarnpkg/fslib`s `copyPromise` does a lot of `lstatPromise` calls that can be skipped.

Ref https://github.com/yarnpkg/berry/issues/4057

**How did you fix it?**

If `copyFolder` created a directory then skip `lstatPromise` calls on all entries in it since we know they don't exist.

**Benchmark results**

On the gatsby benchmark this removes 7976 out of 8090 `statx` syscalls when unplugging.

```sh
YARN_IGNORE_PATH=1 hyperfine -w 1 --prepare="rm -rf .yarn/unplugged"\
  "node ./master.cjs --mode skip-build"\
  "node ./skip-stats.cjs --mode skip-build"
Benchmark 1: node ./master.cjs --mode skip-build
  Time (mean ± σ):      3.843 s ±  0.085 s    [User: 4.785 s, System: 2.640 s]
  Range (min … max):    3.758 s …  4.039 s    10 runs

Benchmark 2: node ./skip-stats.cjs --mode skip-build
  Time (mean ± σ):      3.557 s ±  0.064 s    [User: 4.539 s, System: 2.285 s]
  Range (min … max):    3.510 s …  3.718 s    10 runs

Summary
  'node ./skip-stats.cjs --mode skip-build' ran
    1.08 ± 0.03 times faster than 'node ./master.cjs --mode skip-build'
```

**Checklist**
- [x] I have read the [Contributing Guide](https://yarnpkg.com/advanced/contributing).
- [x] I have set the packages that need to be released for my changes to be effective.
- [x] I will check that all automated PR checks pass before the PR gets reviewed.